Update proj3_weather_poems.py
Updated the file so that the API key can be retrieved from the .env file rather than being put as an input and created an error message if something is wrong with the API key or if the API key is missing. Also removed city=input("Enter a City:") and print(get_weather(city)) since these lines were used for testing if the API call worked.